New updates!

+++ src/main/java/com/chosensolutions/cusbe/domain/dto/FriendDto.java
@@ -2,12 +2,12 @@ package com.chosensolutions.cusbe.domain.dto;
 
 import lombok.Data;
 
-import java.io.Serializable;
+import java.math.BigInteger;
 
 @Data
 public class FriendDto {
 
-    private long userId;
+    private BigInteger userId;
     private String email;
     private String firstName;
     private String lastName;
